+++ frame/evm/precompile/contracts/ed25519/src/lib.rs
@@ -28,10 +28,10 @@ use evm::{ExitError, ExitSucceed};
 pub struct Ed25519Verify;
 
 impl LinearCostPrecompile for Ed25519Verify {
-	const BASE: usize = 15;
-	const WORD: usize = 3;
+	const BASE: u64 = 15;
+	const WORD: u64 = 3;
 
-	fn execute(input: &[u8], _: usize) -> core::result::Result<(ExitSucceed, Vec<u8>), ExitError> {
+	fn execute(input: &[u8], _: u64) -> core::result::Result<(ExitSucceed, Vec<u8>), ExitError> {
 		if input.len() < 128 {
 			return Err(ExitError::Other("input must contain 128 bytes".into()));
 		};
@@ -66,7 +66,7 @@ mod tests {
 	#[test]
 	fn test_empty_input() -> std::result::Result<(), ExitError> {
 		let input: [u8; 0] = [];
-		let cost: usize = 1;
+		let cost: u64 = 1;
 
 		match Ed25519Verify::execute(&input, cost) {
 			Ok((_, _)) => {
@@ -109,7 +109,7 @@ mod tests {
 		input.extend_from_slice(&signature.to_bytes());
 		assert_eq!(input.len(), 128);
 
-		let cost: usize = 1;
+		let cost: u64 = 1;
 
 		match Ed25519Verify::execute(&input, cost) {
 			Ok((_, output)) => {
